fixes twython-related failures
- Splits twitter functionalities in two modules (with and without twython dependencies);
- Fixes some Python 2.6 compatibility problems.
